backend/requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==4.2.5
django-cors-headers==4.2.0
django-js-asset==2.1.0
django-mptt==0.14.0
djangorestframework==3.14.0
drf-spectacular==0.26.4
exceptiongroup==1.1.3
flake8==6.1.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
mccabe==0.7.0
packaging==23.1
Pillow==10.0.1
pluggy==1.3.0
psycopg2-binary==2.9.8
pycodestyle==2.11.0
pyflakes==3.1.0
pytest==7.4.2
pytest-django==4.5.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.10.3
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
uritemplate==4.1.1
